Fix model optional field loading error
If you previously persisted a model with optional fields and you persist
the same model with data that is missing the optional fields, the next
time you load the model, the result will contain the fields persisted in
the previous instance.

Fix this by persisting the most recent field names. Missing fields will
not be loaded.